Move server errors enum to common types file
