+++ packages/webcomponents/src/api/Business.ts
@@ -23,11 +23,6 @@ export enum BusinessType {
   government_entity = 'government_entity',
 }
 
-export enum BusinessFormServerErrors {
-  fetchData = 'Error retrieving business data',
-  patchData = 'Error updating business data'
-}
-
 export interface IAddress {
   id?: string;
   platform_account_id?: string;
