+++ packages/webcomponents/src/components/business-forms/business-form-stepped/legal-address-form/legal-address-form-step.tsx
@@ -1,10 +1,11 @@
 import { Component, Host, Method, Prop, State, h, Event, EventEmitter } from '@stencil/core';
 import { FormController } from '../../../form/form';
 import Api, { IApiResponse } from '../../../../api/Api';
-import { Address, BusinessFormServerErrors, IAddress, IBusiness } from '../../../../api/Business';
+import { Address, IAddress, IBusiness } from '../../../../api/Business';
 import { parseAddressInfo } from '../../utils/payload-parsers';
 import { legalAddressSchema } from '../../schemas/business-legal-address-form-schema';
 import { config } from '../../../../../config';
+import { BusinessFormServerErrors } from '../../utils/business-form-types';
 
 /**
  * @exportedPart label: Label for inputs
